add bash wrapper for enchant spellchecker, important for Finnish
By default, enchant freezes Emacs if used as the "ispell program" in ispell.el,
when any but the default dictionary is used.

This is because enchant uses hunspell-style locale names instead of language
names to choose the dictionary, but ispell.el doesn't know that.

This bash script maps the language name to a locale name, and mogrifies the
command-line arguments to enchant so that e.g. "enchant -a -d finnish" becomes
"enchant -a -d fi_FI".

See ispell-dictionary-base-alist in ispell.el.