flake8: disable E402 (module level import not at top of file)
This warning causes problems in some code when autopep8 "fixes" it without
thinking.

E.g. when using Pygtk, the gi requires must be invoked before we import the
actual GTK libraries, so that the correct version gets loaded.

Also a common pattern is to import logging and configure the logger before
importing other modules.