Fix spellchecker settings for hyphenated compound words in Finnish
Remove the dash ("-") from other-characters (see `ispell-local-dictionary-alist`
in `ispell.el`).

In Finnish, the dash occurs in a compound word basically only when one of the
parts is a foreign word (not yet established as a loanword) or a proper name.
Thus each part should be spell-checked separately, instead of checking the whole
compound as a single word. The whole compound is likely not in the dictionary,
while its parts are likely to be.