experiment with one-key variant of typo zapper
Now both functions are activated by pressing <f8> (once or twice, respectively).
No need to reach for the shift key in normal usage. The <S-f8> binding is still
available, to immediately ask which correction to use, without autozapping first.
This is useful e.g. if you know the first suggestion will be wrong.